backend/pyproject.toml
[project]
name = "food-social-network"
version = "0.1.0"
description = "Food Social Network API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiobotocore==2.21.1",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "dishka[fastapi]>=1.6.0",
    "elasticsearch>=9.0.1",
    "fastapi>=0.109.0",
    "faststream[nats]>=0.5.42",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.25",
    "uvicorn[standard]>=0.27.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.1.13",
    "types-aiobotocore[essential]==2.21.1",
    "uv-sort>=0.5.1",
]
test = [
    "httpx>=0.26.0",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]

[tool.ruff]
target-version = "py311"
exclude = ["venv", ".venv", ".env", "src/typings/external"]

line-length = 120
lint.select = ["ALL"]
lint.ignore = [
    # modules
    "C90", # mccabe complexity
    "DJ",  # django
    "T10", # debugger

    # specific rules
    "COM",    # flake8-commas
    "S104",   # binging to all interfaces
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",   # ignore one-line docstring
    "D203",
    "D212",
    "D213",
    "D400",
    "D415",
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
    "S101",   # prohibited assert statement
    "ARG001", # unused function argument
    "ANN401", # disallow Any type
    "RUF001", # string contrains ambigious o
    "PLR0913", # too many arguments in function signature
]
format.skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["INP001"]
"tests/*" = ["ANN", "INP001"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
exclude = ["typings/external", "typings/external/aiobotocore_s3/type_defs.pyi"]
show_error_context = true
show_absolute_path = true
pretty = true
ignore_missing_imports = true
strict = false
explicit_package_bases = true
